+++ driver/main.cpp
@@ -681,8 +681,22 @@ void registerPredefinedTargetVersions() {
     VersionCondition::addPredefinedGlobalIdent("PPC64");
     registerPredefinedFloatABI("PPC_SoftFloat", "PPC_HardFloat");
     if (triple.getOS() == llvm::Triple::Linux) {
-      VersionCondition::addPredefinedGlobalIdent(
-          triple.getArch() == llvm::Triple::ppc64 ? "ELFv1" : "ELFv2");
+      const llvm::SmallVector<llvm::StringRef> features{};
+      const std::string abi = getABI(triple, features);
+      VersionCondition::addPredefinedGlobalIdent(abi == "elfv1" ? "ELFv1"
+                                                                : "ELFv2");
+      if (global.params.ppcUseIEEE128) {
+        if (triple.getEnvironment() == llvm::Triple::GNU) {
+          // Only GLibc needs this for IEEELongDouble
+          VersionCondition::addPredefinedGlobalIdent("D_PPCUseIEEE128");
+          if (!triple.isLittleEndian()) {
+            warning(Loc(), "float ABI 'ieeelongdouble' is not well-supported "
+                           "on big-endian POWER systems");
+          }
+        } else {
+          warning(Loc(), "float ABI 'ieeelongdouble' is not supported by the target system");
+        }
+      }
     }
     break;
   case llvm::Triple::arm: