+++ gen/target.cpp
@@ -64,9 +64,20 @@ llvm::Type *getRealType(const llvm::Triple &triple) {
   case Triple::wasm64:
     return LLType::getFP128Ty(ctx);
 
+  case Triple::ppc64:
+  case Triple::ppc64le:
+    if (triple.isMusl()) { // Musl uses double
+      return LLType::getDoubleTy(ctx);
+    }
+    // dual-ABI complications: PPC only has IEEE 128-bit quad precision on
+    // Linux, IBM double-double is available on both AIX and Linux.
+    return triple.isOSLinux() && global.params.ppcUseIEEE128
+               ? LLType::getFP128Ty(ctx)
+               : LLType::getPPC_FP128Ty(ctx);
+
   default:
     // 64-bit double precision for all other targets
-    // FIXME: PowerPC, SystemZ, ...
+    // FIXME: SystemZ, ...
     return LLType::getDoubleTy(ctx);
   }
 }
@@ -156,6 +167,7 @@ void Target::_init(const Param &params) {
   const auto IEEEdouble = &APFloat::IEEEdouble();
   const auto x87DoubleExtended = &APFloat::x87DoubleExtended();
   const auto IEEEquad = &APFloat::IEEEquad();
+  const auto PPCDoubleDouble = &APFloat::PPCDoubleDouble();
   bool isOutOfRange = false;
 
   RealProperties.nan = CTFloat::nan;
@@ -197,6 +209,18 @@ void Target::_init(const Param &params) {
     RealProperties.min_exp = -16381;
     RealProperties.max_10_exp = 4932;
     RealProperties.min_10_exp = -4931;
+  } else if (targetRealSemantics == PPCDoubleDouble) {
+    RealProperties.max =
+        CTFloat::parse("0x1.fffffffffffff7ffffffffffff8p1023", isOutOfRange);
+    RealProperties.min_normal = CTFloat::parse("0x1p-969", isOutOfRange);
+    RealProperties.epsilon =
+        CTFloat::parse("0x0.000000000000000000000000008p-969", isOutOfRange);
+    RealProperties.dig = 31;
+    RealProperties.mant_dig = 106;
+    RealProperties.max_exp = 1024;
+    RealProperties.min_exp = -968;
+    RealProperties.max_10_exp = 308;
+    RealProperties.min_10_exp = -291;
   } else {
     // leave initialized with host real_t values
     warning(Loc(), "unknown properties for target `real` type, relying on D "
@@ -240,6 +264,19 @@ const char *TargetCPP::typeMangle(Type *t) {
     // `long double` on Android/x64 is __float128 and mangled as `g`
     bool isAndroidX64 = triple.getEnvironment() == llvm::Triple::Android &&
                         triple.getArch() == llvm::Triple::x86_64;
+    if (triple.getArch() == llvm::Triple::ppc64 ||
+        triple.getArch() == llvm::Triple::ppc64le) {
+      if (global.params.ppcUseIEEE128 &&
+          triple.getEnvironment() == llvm::Triple::GNU) {
+        return "u9__ieee128";
+      }
+      if (size(t) == 16) {
+        // IBM long double
+        return "g";
+      }
+      // fall back to 64-bit double type
+      return "e";
+    }
     return isAndroidX64 ? "g" : "e";
   }
   return nullptr;